refactor user update validation to use a dedicated validation class
